requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
Pillow==10.2.0
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.1
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy==1.26.4
twilio==8.13.0
colorlog==6.8.2
pyautogui==0.9.54
undetected-chromedriver==3.5.5
2captcha-python==1.3.0
PyYAML==6.0.1
cryptography==42.0.4
